+++ Assets/Systems/RenderSys.cs
@@ -8,20 +8,20 @@ public class RenderSys : ISystem
     {
         World.Instance.ForEach<Position>((entity, position) =>
         {
-            ECSManager.Instance.UpdateShapePosition((uint)entity.Id, new Vector2(position.Value.X, position.Value.Y));
+            ECSManager.Instance.UpdateShapePosition(entity.Id, new Vector2(position.Value.X, position.Value.Y));
         });
 
         // Est-ce qu'il faudrait mettre l'update du size dans la boucle précédente ? -> pb de maintenabilité ?
         World.Instance.ForEach<Size>((entity, size) =>
         {
-            ECSManager.Instance.UpdateShapeSize((uint)entity.Id, size.Value.Scale);
+            ECSManager.Instance.UpdateShapeSize(entity.Id, size.Value.Scale);
         });
 
         World.Instance.ForEach<ColorCompo>((entity, colorCompo) =>
         {
             if (colorCompo.HasValue)
             {
-                ECSManager.Instance.UpdateShapeColor((uint)entity.Id, colorCompo.Value.ShapeColor);
+                ECSManager.Instance.UpdateShapeColor(entity.Id, colorCompo.Value.ShapeColor);
             }
         });
     }
